refactor: change findCategoryById method to throw an exception when the id is not found

+++ src/main/java/enzosdev/desafio_junior/service/CategoryService.java
@@ -31,8 +31,9 @@ public class CategoryService {
     }
 
 
-    public Optional<Category> findCategoryById(Long categoryId){
-        return categoryRepository.findById(categoryId);
+    public Category findCategoryById(Long categoryId){
+         return categoryRepository.findById(categoryId)
+                 .orElseThrow(()-> new CategoryNotFoundException("Category Not Found."));
     }
 
 
@@ -44,9 +45,11 @@ public class CategoryService {
 
 
     public void deleteCategoryById(Long id){
-        findCategoryById(id).orElseThrow(()-> new CategoryNotFoundException("Category Not Found"));
+        if (categoryRepository.findById(id).isPresent()){
         categoryRepository.deleteById(id);
-
+        }else{
+            throw new CategoryNotFoundException("Category Not Found.");
+        }
     }
 
 }
